requirements.txt
Django==3.2.9
djangorestframework==3.12.4
Markdown==3.3.6
django-filter==21.1
